move logic to cleanup
