+++ web/src/components/Cleanup.vue
@@ -1,14 +1,12 @@
 <script>
 import { mapMutations } from 'vuex';
-import { CHANGE_AXIS_LABEL, CHANGE_IMPUTATION_OPTIONS } from '@/store/actions.type';
+import { CHANGE_AXIS_LABEL } from '@/store/actions.type';
 import { SET_SELECTION } from '@/store/mutations.type';
 import {
   rowMenuOptions,
   defaultRowOption,
   colMenuOptions,
   defaultColOption,
-  mcar_imputation_methods,
-  mnar_imputation_methods,
 } from '@/utils/constants';
 import { base26Converter } from '@/utils';
 import SaveStatus from '@/components/SaveStatus.vue';
@@ -34,10 +32,6 @@ export default {
       },
       defaultColOption,
       defaultRowOption,
-      settingsDialog: false,
-      mnarImputationMethods: mnar_imputation_methods,
-      mcarImputationMethods: mcar_imputation_methods,
-      imputation: {},
     };
   },
   computed: {
@@ -65,11 +59,31 @@ export default {
       }
       return `Column ${base26Converter(members[0] + 1)}`;
     },
-  },
-  watch: {
-    dataset() {
-      this.imputation.mnar = this.dataset.imputationMNAR;
-      this.imputation.mcar = this.dataset.imputationMCAR;
+    rowHeaders() {
+      return this.dataset.row.labels.map(
+        (rowType, i) => {
+          const header = this.createHeader(rowType, defaultRowOption, i + 1);
+          header.clazz.push(...this.getSelectionClasses('row', i));
+          return header;
+        },
+      );
+    },
+    columns() {
+      const rows = this.dataset.sourcerows;
+      const f = v => (typeof v === 'number' ? v.toFixed(3) : v);
+
+      return this.dataset.column.labels.map((colType, i) => {
+        const column = {
+          index: i,
+          header: this.createHeader(colType, defaultColOption, base26Converter(i + 1)),
+          clazz: [`type-${colType}`],
+          values: rows.map(row => f(row[i])),
+        };
+        const selected = this.getSelectionClasses('column', i);
+        column.clazz.push(...selected);
+        column.header.clazz.push(...selected);
+        return column;
+      });
     },
   },
   methods: {
@@ -80,12 +94,64 @@ export default {
       const changes = ranges.members.map(index => ({ context, index, label }));
       await this.$store.dispatch(CHANGE_AXIS_LABEL, { dataset_id: this.id, changes });
     },
-    async saveImputationSettings() {
-      this.settingsDialog = false;
-      await this.$store.dispatch(CHANGE_IMPUTATION_OPTIONS, {
-        dataset_id: this.id,
-        options: this.imputation,
+
+    createHeader(type, defaultType, text) {
+      const header = {
+        text,
+        clazz: [`type-${type}`],
+      };
+      if (type !== defaultType) {
+        // icon
+        header.text = '';
+        header.clazz.push('mdi', this.$vuetify.icons[type]);
+      }
+      return header;
+    },
+    cellClasses(rowIndex) {
+      const rowType = this.dataset.row.labels[rowIndex];
+      return [`type-${rowType}`, ...this.getSelectionClasses('row', rowIndex)];
+    },
+    getSelectionClasses(axis, index) {
+      if (this.selected.type !== axis) {
+        return [];
+      }
+      const includes = this.selected.ranges.includes(index);
+      const classList = [];
+      if (includes.member) {
+        classList.push('active');
+      }
+      if (includes.first) {
+        classList.push(`${axis}First`);
+      }
+      if (includes.last) {
+        classList.push(`${axis}Last`);
+      }
+      return classList;
+    },
+    onRowClick({ event, rowIndex }) {
+      this.setSelection({
+        key: this.id,
+        event,
+        axis: 'row',
+        idx: rowIndex,
       });
+    },
+    onColumnClick({ event, columnIndex }) {
+      this.setSelection({
+        key: this.id,
+        event,
+        axis: 'column',
+        idx: columnIndex,
+      });
+    },
+    onCellClick({ event, rowIndex, columnIndex }) {
+      const columnType = this.dataset.column.labels[columnIndex];
+      const rowType = this.dataset.row.labels[rowIndex];
+      if (rowType === 'header') {
+        this.onColumnClick(event, columnIndex);
+      } else if (columnType === 'key') {
+        this.onRowClick(event, rowIndex);
+      }
     },
   },
 };
@@ -123,33 +189,10 @@ v-layout.cleanup-wrapper(row)
       v-spacer
 
       save-status
-      v-dialog(v-model="settingsDialog", max-width="500")
-        template(v-slot:activator="{ on }")
-          v-btn(icon, v-on="on")
-            v-icon {{ $vuetify.icons.settings }}
-        v-card
-          v-card-title.headline Imputation Settings
-          v-card-text
-            v-select(
-                item-value="value",
-                item-text="label",
-                :items="mnarImputationMethods",
-                label="MNAR imputation method",
-                v-model="imputation.mnar")
-            v-select(
-                item-value="value",
-                item-text="label",
-                :items="mcarImputationMethods",
-                label="MCAR imputation method",
-                v-model="imputation.mcar")
-          v-card-actions
-            v-spacer
-            v-btn(flat, @click="settingsDialog = false") Close
-            v-btn(flat, @click="saveImputationSettings") Save
-      v-icon {{ $vuetify.icons.download }}
 
-    data-table.cleanup-table(v-bind="{ id, dataset, selected }",
-        @setselection="setSelection")
+    data-table.cleanup-table(:row-headers="rowHeaders", :columns="columns",
+        :cellClasses="cellClasses", @row-click="onRowClick($event)",
+        @column-click="onColumnClick($event)", @cell-click="onCellClick($event)")
 </template>
 
 <style lang="scss">
