+++ web/src/components/DataTable.vue
@@ -20,116 +20,36 @@
 import 'vue-virtual-scroller/dist/vue-virtual-scroller.css';
 import { RecycleScroller } from 'vue-virtual-scroller';
 
-import { base26Converter } from '../utils';
-import { defaultRowOption, defaultColOption } from '../utils/constants';
-
 export default {
   components: {
     RecycleScroller,
   },
   props: {
-    dataset: {
-      type: Object,
+    rowHeaders: { // {text: string, clazz: string[]}[]
+      type: Array,
       required: true,
     },
-    id: {
-      type: String,
+    columns: { // {index: number, header: ..., clazz: string[], values: string[]}[]
+      type: Array,
       required: true,
     },
-    selected: {
-      type: Object,
+    cellClasses: { // (rowIndex: number) => string[]
+      type: Function,
       required: true,
-    },
-  },
-  computed: {
-    rowHeaders() {
-      return this.dataset.row.labels.map(
-        (rowType, i) => {
-          const header = this.createHeader(rowType, defaultRowOption, i + 1);
-          header.clazz.push(...this.getSelectionClasses('row', i));
-          return header;
-        },
-      );
-    },
-    columns() {
-      const rows = this.dataset.sourcerows;
-      const f = v => (typeof v === 'number' ? v.toFixed(3) : v);
-
-      return this.dataset.column.labels.map((colType, i) => {
-        const column = {
-          index: i,
-          header: this.createHeader(colType, defaultColOption, base26Converter(i + 1)),
-          clazz: [`type-${colType}`],
-          values: rows.map(row => f(row[i])),
-        };
-        const selected = this.getSelectionClasses('column', i);
-        column.clazz.push(...selected);
-        column.header.clazz.push(...selected);
-        return column;
-      });
     },
   },
   methods: {
-    createHeader(type, defaultType, text) {
-      const header = {
-        text,
-        clazz: [`type-${type}`],
-      };
-      if (type !== defaultType) {
-        // icon
-        header.text = '';
-        header.clazz.push('mdi', this.$vuetify.icons[type]);
-      }
-      return header;
-    },
-    cellClasses(rowIndex) {
-      const rowType = this.dataset.row.labels[rowIndex];
-      return [`type-${rowType}`, ...this.getSelectionClasses('row', rowIndex)];
-    },
     setSelection(selection) {
       this.$emit('setselection', selection);
     },
-    getSelectionClasses(axis, index) {
-      if (this.selected.type !== axis) {
-        return [];
-      }
-      const includes = this.selected.ranges.includes(index);
-      const classList = [];
-      if (includes.member) {
-        classList.push('active');
-      }
-      if (includes.first) {
-        classList.push(`${axis}First`);
-      }
-      if (includes.last) {
-        classList.push(`${axis}Last`);
-      }
-      return classList;
-    },
     onRowClick(event, rowIndex) {
-      this.setSelection({
-        key: this.id,
-        event,
-        axis: 'row',
-        idx: rowIndex,
-      });
+      this.$emit('row-click', { event, rowIndex });
     },
     onColumnClick(event, columnIndex) {
-      this.setSelection({
-        key: this.id,
-        event,
-        axis: 'column',
-        idx: columnIndex,
-      });
+      this.$emit('column-click', { event, columnIndex });
     },
     onCellClick(event, rowIndex, columnIndex) {
-      const columnType = this.dataset.column.labels[columnIndex];
-      const rowType = this.dataset.row.labels[rowIndex];
-      if (rowType === 'header') {
-        this.onColumnClick(event, columnIndex);
-      } else if (columnType === 'key') {
-        this.onRowClick(event, rowIndex);
-      }
+      this.$emit('cell-click', { event, rowIndex, columnIndex });
     },
   },
 };
